pyproject.toml
[project]
name = "finsynapse"
version = "0.1.0"
description = "Personal multi-market macro data pipeline + market thermometer + local LLM insights"
authors = [
    { name = "hg.dendi", email = "hg.dendi@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "akshare>=1.18.59",
    "beautifulsoup4>=4.14.3",
    "jinja2>=3.1.6",
    "lxml>=6.1.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "pyarrow>=24.0.0",
    "pydantic-settings>=2.14.0",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "streamlit>=1.57.0",
    "typer>=0.25.0",
    "yfinance>=1.3.0",
]

[project.scripts]
finsynapse = "finsynapse.cli:app"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-vcr>=1.0.2",
    "vcrpy>=8.1.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
